+++ e2e/node/utils/mock-replica.ts
@@ -6,7 +6,6 @@ import type {
   Cert,
   Nonce,
   RequestId,
-  ReadStateResponse,
   HashTree,
   Signed,
   UnSigned,
@@ -400,7 +399,7 @@ export async function prepareV3ReadStateResponse({
     signature,
     delegation,
   };
-  const responseBody: ReadStateResponse = {
+  const responseBody = {
     certificate: Cbor.encode(cert),
   };
 
@@ -441,7 +440,7 @@ export async function prepareV3ReadStateRootSubnetResponse({
     tree,
     signature,
   };
-  const responseBody: ReadStateResponse = {
+  const responseBody = {
     certificate: Cbor.encode(cert),
   };